notes/backend/requirement.txt
mysqlclient==1.4.6
django==2.2.3
Jinja2==2.10.3
Markdown==3.1.1
pygments==2.5.2
coreapi==2.3.3
dj-rest-auth==1.1.1
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.11.0
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
pillow==7.0.0
crypto==1.4.1
pycryptodome==3.9.7
